estrutura finalizada, restando apenas ajustes de estilo

+++ database/migrations/2021_07_10_003622_create_clients_table.php
@@ -19,8 +19,7 @@ class CreateClientsTable extends Migration
             $table->string('cpf', 14);
             $table->date('data_nasc');
             $table->date('data_cadastro');
-            // $table->decimal('renda', 5,2);
-            $table->integer('renda');
+            $table->decimal('renda', 15,2);
 
             $table->timestamps();
         });
